- Fixed date to string JS function in input.vm. Numbers smaller than 10 are now prefixed with "0".
- Fixed NPE in UserDelegation in case admin groups is null.

- Fixed CustomApprovalUtil to support status that is set to allow all
transitions. The logic to retrieve the required ActionDescriptor has
been updated. In the future initialize post function needs to be updated
to allow user to specify which transition to use, in case where there
are more than one.

- Fixed bug in approval logic that when transition failed, Approval Data
is still updated. Now Approval Data is only updated when no transition
is reuqired or after a successful transition.