- Version 1.8.1.
- Fixed bug where CustomApprovalIssueEventListener will turn on the
Issue Updated Event Indicator by itself, because calculated request
participant hasn't changed. The logic to turn the flag off requires
request participant to be changed, and so the flag isn't turned off
properly and must wait for request participant to change again. The fix
is to check the value of request participant, and only update the issue
if it is changed.